made further changes to the notification system and made functions to do calls to the server which will not work atm. made minor cosmetic changes.

+++ lib/notifications/followed_stock_list_view.dart
@@ -2,12 +2,14 @@ import 'dart:collection';
 import 'dart:math';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_application_2/notifications/followed_stock_model.dart';
 import 'package:provider/provider.dart';
 import '../model/main_model.dart';
 import '../pages/stock_page.dart';
+import 'package:http/http.dart' as http;
 import '../followed_stock_model.dart';
+import '../services/api_service.dart';
 
 Future<void> DeleteAllFavorites() async {
   final user = FirebaseAuth.instance.currentUser;
@@ -291,6 +293,10 @@ void createNotification(BuildContext context, String selectedButton, String stoc
   final db = FirebaseFirestore.instance;
   final userID = FirebaseAuth.instance.currentUser?.uid;
 
+  // Retrieve FCM token
+  FirebaseMessaging messaging = FirebaseMessaging.instance;
+  String? fcmToken = await messaging.getToken();
+
   // Check the number of existing notifications for the user
   final notificationSnapshot = await db
       .collection("notifications")
@@ -321,45 +327,20 @@ void createNotification(BuildContext context, String selectedButton, String stoc
     return; // Exit the function if the maximum notification count is reached
   }
 
-  // Check if the user already has a notification for the stock
-  final stockSnapshot = await db
-      .collection("notifications")
-      .where("userID", isEqualTo: userID)
-      .where("tickerID", isEqualTo: stockName)
-      .get();
-  if (stockSnapshot.docs.isNotEmpty) {
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text("Existing Notification"),
-          content: const Text("You already have a notification for this stock."),
-          actions: [
-            TextButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-              child: const Text("OK"),
-            ),
-          ],
-        );
-      },
-    );
-    return; // Exit the function if the user already has a notification for the stock
-  }
+  // Make the first POST call to send user ID and FCM token
+  ApiService.sendUserData(userID, fcmToken);
 
-  // Add a new document with a generated id
-  final data = {
-    "userID": userID,
-    "percentage": selectedButton,
-    "tickerID": stockName,
-  };
-  db.collection("notifications").add(data);
+  // Make the second POST call to send user ID, percentage, and ticker ID
+  ApiService.sendNotificationData(userID, selectedButton, stockName);
+
+  // Close the popup page
+  Navigator.of(context).pop();
 }
 
 Future<void> deleteNotification(String stockCode) async {
   final user = FirebaseAuth.instance.currentUser;
   final userID = user?.uid;
+
   var db = FirebaseFirestore.instance;
   final snapshot = await db
       .collection("notifications")
@@ -367,10 +348,21 @@ Future<void> deleteNotification(String stockCode) async {
       .where("tickerID", isEqualTo: stockCode)
       .get();
 
-  for (final doc in snapshot.docs) {
-    await doc.reference.delete();
+  // Check if the document exists
+  if (snapshot.docs.isNotEmpty) {
+    // Get the percentage value from the document
+    final percentage = snapshot.docs.first['percentage'];
+
+    // Delete notification in the API service
+    await ApiService.deleteNotification(userID, stockCode, percentage);
+
+    // Delete notification in Firestore
+    for (final doc in snapshot.docs) {
+      await doc.reference.delete();
+    }
   }
 }
+
 
 class NotificationsPage extends StatefulWidget {
   const NotificationsPage({Key? key}) : super(key: key);
@@ -380,7 +372,7 @@ class NotificationsPage extends StatefulWidget {
 }
 
 class _NotificationsPageState extends State<NotificationsPage> {
-  Future<List<MainModel>> getNotificationStocks() async {
+  Future<List<Map<String, dynamic>>> getNotificationStocks() async {
     final user = FirebaseAuth.instance.currentUser;
     final userID = user?.uid;
 
@@ -390,14 +382,13 @@ class _NotificationsPageState extends State<NotificationsPage> {
           .where('userID', isEqualTo: userID)
           .get();
 
-      final notifications = snapshot.docs.map((doc) => doc['tickerID'])
-          .toList();
-
-      final mainData = MainModel.data;
-
-      final notificationStocks = mainData.values
-          .where((stock) => notifications.contains(stock.name))
-          .toList();
+      final notificationStocks = snapshot.docs.map((doc) {
+        final data = doc.data();
+        return {
+          'name': data['tickerID'],
+          'percentage': data['percentage'],
+        };
+      }).toList();
 
       return notificationStocks;
     }
@@ -412,7 +403,7 @@ class _NotificationsPageState extends State<NotificationsPage> {
         // Refresh the notification stocks
         setState(() {});
       },
-      child: FutureBuilder<List<MainModel>>(
+      child: FutureBuilder<List<Map<String, dynamic>>>(
         future: getNotificationStocks(),
         builder: (context, snapshot) {
           if (snapshot.connectionState == ConnectionState.waiting) {
@@ -435,27 +426,46 @@ class _NotificationsPageState extends State<NotificationsPage> {
                   padding: const EdgeInsets.all(10),
                   child: Text(
                     "You have $remainingNotifications unique notifications left",
-                    style: TextStyle(
+                    style: const TextStyle(
                       fontSize: 20, // Increase the font size
                     ),
                   ),
                 ),
+                const Divider(), // Add a separating line
                 Expanded(
                   child: ListView.builder(
                     itemCount: notificationStocks?.length,
                     itemBuilder: (context, index) {
                       final notificationStock = notificationStocks![index];
+                      final stockName = notificationStock['name'];
+                      final percentage = notificationStock['percentage'];
+
                       return ListTile(
-                        title: Text(notificationStock.name),
-                        subtitle: Text(notificationStock.name),
+                        title: Row(
+                          children: [
+                            Text(
+                              stockName,
+                              style: const TextStyle(
+                                fontWeight: FontWeight.bold,
+                              ),
+                            ),
+                            const SizedBox(width: 10),
+                            // Add spacing between stock name and percentage
+                            Text(
+                              "$percentage%",
+                              style: const TextStyle(
+                                color: Colors.blue,
+                              ),
+                            ),
+                          ],
+                        ),
                         trailing: Row(
                           mainAxisSize: MainAxisSize.min,
                           children: [
                             IconButton(
                               onPressed: () async {
                                 // Handle trash button functionality
-                                await deleteNotification(
-                                    notificationStock.name);
+                                await deleteNotification(stockName);
                                 setState(() {});
                               },
                               icon: const Icon(Icons.delete),
@@ -466,8 +476,8 @@ class _NotificationsPageState extends State<NotificationsPage> {
                           Navigator.of(context).push(MaterialPageRoute(
                             builder: (BuildContext context) =>
                                 StockPage(
-                                  stockCode: notificationStock.name,
-                                  stockName: notificationStock.name,
+                                  stockCode: stockName,
+                                  stockName: stockName,
                                 ),
                           ));
                         },
@@ -483,4 +493,5 @@ class _NotificationsPageState extends State<NotificationsPage> {
     );
   }
 }
+
